Move middleware outside resources
