+++ backend/functions/src/api/middleware/error-handler.ts
