+++ backend/functions/src/api/middleware/logger.ts
